requirements/base.txt
# core packages
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
drf-yasg==1.21.6
django-filter==23.2
python-dotenv
django-autoslug==1.9.9
pytils==0.4.1
django-cors-headers==4.1.0
Pillow==9.5.0
argon2-cffi==21.3.0
redis==4.6.0
celery==5.3.1
flower==2.0.0
django-celery-email==3.0.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
coreapi==2.3.3